fix: moved images to assets
