+++ app/page.tsx
@@ -1,4 +1,4 @@
-import logo from '../images/logo.svg';
+import logo from '@/assets/images/logo.svg';
 import Image from 'next/image';
 import { LoginForm } from '@/components/forms/LoginForm';
 import { getAuthServerSession } from '@/lib/authenticate';
